requirements.txt
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
joblib>=1.3
dvc[gdrive]>=2.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.1
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.4.0
billiard==4.2.2
blinker==1.9.0
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.71.0
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.8
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.120.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.42.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gto==1.9.0
gunicorn==23.0.0
h11==0.16.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
multidict==6.7.0
networkx==3.5
numpy==2.3.4
omegaconf==2.3.0
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydot==4.0.1
pygit2==1.19.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.16.3
scmrepo==3.5.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.49.1
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.14
Werkzeug==3.1.3
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
